fix: comprehensive code quality review — fix 56 issues across all drivers Two-round review covering all source files. Key fixes: Round 1 (32 issues): - PID: fix position derivative using e(k-2) instead of e(k-1), fix   incremental PID history indices for D-on-PV formula - Main: sprintf overflow, signed/unsigned dc type, uninitialized vars,   dead code removal, LCD string truncation, K230 counter overflow - Linefollow: dc_base uint8→int16, line-lost gradual deceleration,   magic numbers replaced with named constants - MPU6050: I2C bus recovery on timeout, calibration uses actual good   count, remove incorrect volatile, safe uint16 shift - LCD: INT_MIN negation UB fix, SLPOUT 120ms delay, char bounds check,   COLOR565/GRAY2RGB16 macro parenthesization, cursor off-by-one - Headers: reserved _IDENTIFIER include guards → IDENTIFIER_ pattern - ISR: remove redundant clearInterruptStatus (IIDX auto-clears) - Menu: count=0 guard, dedup init/switch, scroll delay fix Round 2 (24 issues): - UART: fix volatile double-read race in available() and read_byte()   that could return values exceeding buffer size or inject 0x00 bytes - MPU6050: add NACK/error detection after I2C transfers, check all   write_reg returns in init(), gz offset int32 to prevent overflow,   DLPF settling delay before calibration, fail on <50% good samples,   who_am_i returns 0xFF on failure, yaw range normalization ±3600° - Encoder: move static locals to file scope for proper reset, add   int32→int16 speed delta clamping - Key: fix debounce edge loss when scan runs twice before check - Linefollow: saturate g_lost_frames to prevent uint8 wraparound,   remove unused dc_float, deduplicate CCD init in demo - Odometer: saturating pulse_total add, safe INT32_MIN abs via   unsigned, sprintf→snprintf - Motor: brake CC value PERIOD+1 for guaranteed 100% duty - LCD: draw_frame uses dx-1/dy-1 for correct corners, show_str   uint8→uint16 counter with x-bounds check, fix temp>>=1 indent,   GRAY2RGB16_hl macro overflow fix - K230: is_number supports negative sign, limits digit count to 5 - CCD: uint8→uint16 loop counter for future-proofing - Main: rect clamp order (min-size before boundary), negative K230   coord guard, fmodf negative yaw +360 normalization, remaining   sprintf→snprintf, remove unused redraw variable Also adds tsp_linefollow, tsp_odometer, tsp_pid source files to the IAR .ewp project (were missing, causing 11 linker errors). Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>

+++ empty_mspm0g3519/TSP3519/tsp_ccd.c
@@ -97,7 +97,7 @@ void tsp_ccd_set_exposure(uint8_t ms)
 
 void tsp_ccd_flush(uint8_t ccd_id)
 {
-    uint8_t i;
+    uint16_t i;
 
     /* ─── Start integration cycle ─── */
     ccd_si_high(ccd_id);
@@ -124,7 +124,7 @@ void tsp_ccd_flush(uint8_t ccd_id)
 
 void tsp_ccd_snapshot(uint8_t ccd_id, ccd_data_t pixels)
 {
-    uint8_t          i;
+    uint16_t         i;
     DL_ADC12_MEM_IDX mem_idx = ccd_adc_mem_idx(ccd_id);
 
     /* 1. Flush previously integrated frame */
